% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_matrix.R
\name{compute_loglik_matrix}
\alias{compute_loglik_matrix}
\title{Computes n by K log-likelihood matrix for a ud fit object}
\usage{
compute_loglik_matrix(fit, version = c("Rcpp", "R"))
}
\description{
Computes n by K log-likelihood matrix for a ud fit object
}
