% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{compute_loglik_penalized}
\alias{compute_loglik_penalized}
\title{Function to compute penalized log-likelihood.}
\usage{
compute_loglik_penalized(loglik, logplt)
}
\arguments{
\item{loglik:}{the log-likelihood calculated on fitted mixture model.}

\item{logplt:}{log of the penality}
}
\description{
Function to compute penalized log-likelihood.
}
