% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{compute_penalty}
\alias{compute_penalty}
\title{Function to compute log-penalty for one prior covariance in the transformed
space if the transformation is applied.}
\usage{
compute_penalty(U, sigma2, n0 = 0, lambda = 0, S0, alpha = 0.5)
}
\arguments{
\item{U:}{prior covariance matrix (in the transformed space). 
U has to be unconstrained covariance, otherwise makes no sense to penalize.}

\item{sigma2:}{The scalar attached to U.}

\item{n0:}{the penalty strength of inverse-Wishart prior used in ED updates.}

\item{lambda:}{the penalty strength of nuclear norm function used in TED updates.}

\item{S0:}{a positive definite matrix used as the parameter of inverse-Wishart distribution.}

\item{alpha:}{a tuning parameter used in nuclear norm penalty function. 
Default of 0.5 is recommended.}
}
\description{
Function to compute log-penalty for one prior covariance in the transformed
space if the transformation is applied.
}
