% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{create_rank1_matrix_struct}
\alias{create_rank1_matrix_struct}
\title{Function to create the data structure for rank1 matrix}
\usage{
create_rank1_matrix_struct(X, U_rank1)
}
\description{
Function to create the data structure for rank1 matrix
}
