% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{create_scaled_matrix_struct}
\alias{create_scaled_matrix_struct}
\title{Function to create the data structure for scaled matrix}
\usage{
create_scaled_matrix_struct(X, U_scaled)
}
\description{
Function to create the data structure for scaled matrix
}
