% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{create_unconstrained_matrix_struct}
\alias{create_unconstrained_matrix_struct}
\title{Function to create the data structure for unconstrained matrix}
\usage{
create_unconstrained_matrix_struct(X, U_unconstrained)
}
\description{
Function to create the data structure for unconstrained matrix
}
