% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_autocv.R
\name{get_best_fit}
\alias{get_best_fit}
\title{Function to get the best fit on the whole dataset based on cv results.}
\usage{
get_best_fit(X, V, best_fit_cv, control = list(), verbose)
}
\arguments{
\item{X:}{n by R data matrix}

\item{V:}{R by R residual covariance matrix}

\item{best_fit_cv:}{An object storing the best cv results and its parameter configuration.}
}
\value{
best_fit: the fit object on whole dataset.
}
\description{
Function to get the best fit on the whole dataset based on cv results.
}
