% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_autocv.R
\name{get_best_fit_cv}
\alias{get_best_fit_cv}
\title{Function to get the best_fit_cv object based on highest average test log-likelihood.}
\usage{
get_best_fit_cv(res)
}
\arguments{
\item{res:}{the result from ud_fit_cv().}
}
\value{
best_fit_cv: an object storing training results, its corresponding average 
test loglikelihood and parameter configuration.
}
\description{
Function to get the best_fit_cv object based on highest average test log-likelihood.
}
