% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_mat_Q}
\alias{get_mat_Q}
\title{Function to calculate matrix Q where U=QQ^T when U is rank-deficient.}
\usage{
get_mat_Q(U, r)
}
\arguments{
\item{U}{is the rank-deficient matrix to decompose}

\item{r}{is the rank of U.}
}
\description{
Function to calculate matrix Q where U=QQ^T when U is rank-deficient.
}
