% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted.R
\name{get_S}
\alias{get_S}
\title{Function to calculate weighted empirical covariance matrix. Each data point 
is weighted by its posterior weight for a certain component.}
\usage{
get_S(X, p)
}
\arguments{
\item{X:}{data matrix of size $n$ by $R$.}

\item{p:}{a vector of size $n$, containing the posterior weight for a certain 
component for each observation.}
}
\description{
Function to calculate weighted empirical covariance matrix. Each data point 
is weighted by its posterior weight for a certain component.
}
