% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted.R
\name{get_sigma2}
\alias{get_sigma2}
\title{Function to update sigma2 in M-step optimization given 
all the eigenvalues of U. See eq.(82) in the writeup.}
\usage{
get_sigma2(alpha, eigenvals)
}
\arguments{
\item{alpha:}{a number that controls the trade-off between the two nuclear norm terms.}

\item{eigenvals:}{the eigenvalues of U.}
}
\description{
Function to update sigma2 in M-step optimization given 
all the eigenvalues of U. See eq.(82) in the writeup.
}
