% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_Ulist}
\alias{get_Ulist}
\title{Extract the list of fitted U from udr fit object}
\usage{
get_Ulist(fit)
}
\arguments{
\item{fit}{udr object}
}
\description{
Extract the list of fitted U from udr fit object
}
