% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted.R
\name{grad_loglik_per_eigenval}
\alias{grad_loglik_per_eigenval}
\title{Function to calculate the partial derivative w.r.t. one eigenvalue
of U. See equation (80) in https://www.overleaf.com/read/vrgwpskkhbpj.}
\usage{
grad_loglik_per_eigenval(val, p, d, lambda, alpha, sigma2)
}
\arguments{
\item{val:}{the ith eigenvalue that we want to solve for}

\item{d:}{the ith eigenvalue of weighted empirical covariance matrix.}

\item{p:}{the weight vector for a component}

\item{lambda:}{the strength of penalty}

\item{alpha:}{control the trade-off between the two nuclear norm terms.}

\item{sigma2:}{the scalar on U.}
}
\description{
Function to calculate the partial derivative w.r.t. one eigenvalue
of U. See equation (80) in https://www.overleaf.com/read/vrgwpskkhbpj.
}
