% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{logLik.ud_fit}
\alias{logLik.ud_fit}
\title{Title Goes Here}
\usage{
\method{logLik}{ud_fit}(object, version = c("Rcpp", "R"), ...)
}
\arguments{
\item{object}{Describe input argument "fit" here.}

\item{version}{Description of input argument "version" here.}

\item{\dots}{Additional arguments (unused).}
}
\value{
Describe the return value here.
}
\description{
Description goes here.
}
