% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted.R
\name{loglik_neg_per_eigenval}
\alias{loglik_neg_per_eigenval}
\title{The following functions implement nuclear norm penalty to covariance estimation
under scaled model, x_j\sim N(0, \sigma^2\tilde U + I). 
The nuclear norm penalty is specified on \tilde U. 
Function to compute the negative of objective function to be minimized.}
\usage{
loglik_neg_per_eigenval(val, p, d, lambda, alpha, sigma2)
}
\arguments{
\item{val:}{the ith eigenvalue that we want to solve for}

\item{d:}{the ith eigenvalue of weighted empirical covariance matrix.}

\item{p:}{the weight vector for a component}

\item{lambda:}{the strength of penalty}

\item{alpha:}{control the trade-off between the two nuclear norm terms.}

\item{sigma2:}{the scalar on U.}
}
\description{
The following functions implement nuclear norm penalty to covariance estimation
under scaled model, x_j\sim N(0, \sigma^2\tilde U + I). 
The nuclear norm penalty is specified on \tilde U. 
Function to compute the negative of objective function to be minimized.
}
