% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvebnm.R
\name{mvebnm}
\alias{mvebnm}
\title{Title Goes Here.}
\usage{
mvebnm(X, w, U, S, eps, maxiter, tol = 1e-06, verbose = TRUE)
}
\arguments{
\item{X}{Describe input argument "X" here.}

\item{w}{Describe input argument "U" here.}

\item{S}{Describe input argument "S" here.}
}
\description{
Description goes here.
}
\references{
Add references here.
}
