% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{simulate_ebmn_data}
\alias{simulate_ebmn_data}
\title{Function to simulate data from EBNM model for one component.
It first simulate the true mean and then observed data.}
\usage{
simulate_ebmn_data(n, U, V)
}
\arguments{
\item{n}{Number of data points to simulate.}

\item{U}{Prior covariance matrix}

\item{V}{Residual covariance matrix}
}
\description{
Function to simulate data from EBNM model for one component.
It first simulate the true mean and then observed data.
}
