% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{simulate_ebmvnm_data}
\alias{simulate_ebmvnm_data}
\title{Simulate Data from Multivariate Normal Means Model}
\usage{
simulate_ebmvnm_data(n, w, U, S = diag())
}
\arguments{
\item{w}{A numeric vector of length k specifying the prior mixture
weights. All entries must be non-negative, but need not sum to 1;
the mixture weights are automatically normalized to sum to 1.}

\item{U}{A list of length k specifying the covariance matrices in
the mixture-of-multivariate-normals prior; list element
\code{U[[i]]} is the m x m covariance matrix for the ith mixture
component.}

\item{S}{The m x m residual covariance matrix.}
}
\value{
Describe output here.
}
\description{
Simulate n data points from the multivariate normal
  means model. See \code{\link{mvebnm}} for the model definition.
}
\seealso{
\code{\link{mvebnm}}
}
