% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ud_fit}
\alias{summary.ud_fit}
\alias{print.summary.ud_fit}
\title{Summarize Ultimate Deconvolution Fit}
\usage{
\method{summary}{ud_fit}(object, ...)

\method{print}{summary.ud_fit}(x, ...)
}
\arguments{
\item{object}{An object of class \dQuote{ud_fit}, usually the
result of calling function \code{link{ud_fit}}.}

\item{\dots}{Additional arguments passed to the generic \code{summary}
or \code{print.summary} method.}

\item{x}{An object of class \dQuote{summary.ud_fit}, usually the
result of a call to \code{summary.ud_fit}.}
}
\value{
\code{summary.ud_fit} returns a list of statistics
  summarizing the model fit.
}
\description{
\code{summary} method for the \dQuote{ud_fit} class.
}
