% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted.R
\name{ted_penalized}
\alias{ted_penalized}
\title{Function to regularize U by nuclear penalty from Chi and Lange (2014).
Here I assume V_j = I.}
\usage{
ted_penalized(X, S, p, lambda, alpha, sigma2)
}
\arguments{
\item{X:}{data matrix of size $n$ by $R$.}

\item{S:}{weighted empirical covariance matrix}

\item{p:}{a vector of size $n$, containing the posterior weight for a certain 
component for each observation.}

\item{lambda:}{the strength of penalty}

\item{alpha:}{a number that controls the trade-off between the two nuclear norm terms.}

\item{sigma2:}{the scalar on U.}
}
\description{
Function to regularize U by nuclear penalty from Chi and Lange (2014).
Here I assume V_j = I.
}
