% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted.R
\name{ted_penalized_iw}
\alias{ted_penalized_iw}
\title{Function to regularize U by IW penalty. Here I assume V_j = I.}
\usage{
ted_penalized_iw(X, S, p, lambda, sigma2)
}
\description{
Function to regularize U by IW penalty. Here I assume V_j = I.
}
