% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/posterior.R, R/updates.R
\name{ud_fit_em}
\alias{ud_fit_em}
\alias{compute_posterior_probs}
\alias{update_mixture_weights_em}
\alias{update_resid_covariance}
\alias{update_prior_covariances}
\alias{assign_prior_covariance_updates}
\title{Low-Level Interface for Fitting Ultimate Deconvolution Models}
\usage{
ud_fit_em(
  fit,
  covupdates = rep("none", length(fit$U)),
  control = list(),
  verbose = TRUE
)

compute_posterior_probs(fit, version = c("Rcpp", "R"))

update_mixture_weights_em(fit, update = c("em", "none"))

update_resid_covariance(
  fit,
  update = c("em", "none"),
  version = c("Rcpp", "R")
)

update_prior_covariances(
  fit,
  covupdates = rep("none", length(fit$U)),
  minval = 1e-14
)

assign_prior_covariance_updates(covtypes, control)
}
\arguments{
\item{fit}{An Ultimate Deconvolution model fit. Typically,
this will be an output of \code{\link{ud_init}} or \code{ud_fit}.}

\item{covupdates}{Describe input argument "covupdates" here.}

\item{control}{A list of parameters controlling the behaviour of
the model fitting and initialization. See \sQuote{Details}.}

\item{verbose}{When \code{verbose = TRUE}, information about the
algorithm's progress is printed to the console at each
iteration.}

\item{version}{Describe input argument "version" here.}

\item{update}{Describe input argument "update" here.}

\item{minval}{Describe input argument "minval" here.}

\item{covtypes}{Describe input argument "covtypes" here.}
}
\value{
An Ultimate Deconvolution model fit. See
  \code{\link{ud_fit}} for details.
}
\description{
Here we describe the low-level model fitting interface.
}
\details{
Add details here.
}
\examples{
# Add examples here.

}
\seealso{
\code{\link{ud_init}}, \code{\link{ud_fit}}
}
\keyword{internal}
