% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ud_init}
\alias{ud_init}
\title{Function Title Goes Here}
\usage{
ud_init(X, V = diag(ncol(X)), U, w, k)
}
\arguments{
\item{X}{The n x m data matrix, in which each row is an
m-dimensional observation.}

\item{V}{An m x m matrix giving the initial estimate of the
residual covariance matrix.}

\item{U}{A list of length k giving initial estimates of the
covariance matrices in the mixture-of-multivariate-normals prior;
list element \code{U[[i]]} is the m x m covariance matrix for the
ith mixture component. If not provided, the initial estimates are
randomly generated by computing sample covariances on random
subsets of the data \code{X}.}

\item{w}{A numeric vector of length k giving initial estimates of
the prior mixture weights. All entries must be non-negative, but
need not sum to 1; the mixture weights are automatically normalized
to sum to 1. If not provided, the mixture weights are set to
uniform.}

\item{k}{An integer 2 or greater specifying the number of
components in the mixture-of-normals prior. This only needs to be
provided if neither \code{U} nor \code{w} are provided.}
}
\description{
Description of function goes here.
}
\examples{

library(mvtnorm)
set.seed(1)
V <- diag(2) + 1
X <- rmvt(1000,V,df = 4)
U <- list(fixed  = matrix(1,2,2),
          scaled = diag(2),
          rank1  = tcrossprod(c(1,2)),
          ed     = crossprod(matrix(rnorm(4),2,2)),
          teem   = crossprod(matrix(rnorm(4),2,2)))
attr(U$fixed,"covtype")  <- "fixed"
attr(U$scaled,"covtype") <- "scaled"
attr(U$rank1,"covtype")  <- "rank1"
attr(U$ed,"covtype")     <- "ed"
attr(U$teem,"covtype")   <- "teem"
fit0 <- ud_init(X,V,U = U)

}
