% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udr-package.R
\docType{package}
\name{udr-package}
\alias{udr}
\alias{udr-package}
\title{udr: Ultimate Deconvolution for Multivariate Normal Means}
\description{
Implements algorithms for solving the
    multivariate normal means problem via empirical Bayes. The ``Ultimate Deconvolution" name comes its connection to the 
    "Extreme Deconvolution" method \doi{10.1214/10-AOAS439}. The main function is \link{ud_fit}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenslab/udr}
  \item Report bugs at \url{https://github.com/stephenslab/udr/issues}
}

}
\author{
\strong{Maintainer}: Peter Carbonetto \email{peter.carbonetto@gmail.com}

Authors:
\itemize{
  \item Yunqi Yang
  \item Matthew Stephens
}

Other contributors:
\itemize{
  \item Yuxin Zou [contributor]
  \item David Gerard [contributor]
  \item Gao Wang [contributor]
}

}
\keyword{internal}
