% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{update_prior_covariance_rank1_fa_iid}
\alias{update_prior_covariance_rank1_fa_iid}
\title{Perform an M-step update for rank1 prior covariance matrix U 
using factor analyzer in the special case of V_j = I.}
\usage{
update_prior_covariance_rank1_fa_iid(X, u, p)
}
\arguments{
\item{X}{contains observed data of size n \times r.}

\item{u}{the estimate of vector that forms the rank1 U in previous iteration}

\item{p}{is a vector of weights}
}
\description{
Perform an M-step update for rank1 prior covariance matrix U 
using factor analyzer in the special case of V_j = I.
}
