% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{update_prior_covariance_scaled_fa_general}
\alias{update_prior_covariance_scaled_fa_general}
\title{Perform an M-step update for estimating the scalar for prior covariance matrix U0
in the general case where V_j can vary for different observations. U0 can be rank-deficient.}
\usage{
update_prior_covariance_scaled_fa_general(X, U0, V, p, s, r)
}
\arguments{
\item{X}{contains observed data of size n \times r.}

\item{U0}{A known canonical covariance of size r \times r.}

\item{V}{is a 3-d array, in which V[,,j] is the covariance matrix}

\item{p}{is a vector of weights}

\item{s}{is the scalar estimate in previous iteration}

\item{r}{is the rank of U0}
}
\description{
Perform an M-step update for estimating the scalar for prior covariance matrix U0
in the general case where V_j can vary for different observations. U0 can be rank-deficient.
}
