% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R, R/update_prior_subproblems.R
\name{update_prior_covariance_scaled_fa_iid}
\alias{update_prior_covariance_scaled_fa_iid}
\title{Perform an M-step update for estimating the scalar for prior covariance matrix U0
in the special case of V_j = I. U0 can be rank-deficient.}
\usage{
update_prior_covariance_scaled_fa_iid(X, U0, p, s, r)

update_prior_covariance_scaled_fa_iid(X, U0, p, s, r)
}
\arguments{
\item{X}{contains observed data of size n \times r.}

\item{U0}{A known canonical covariance of size r \times r.}

\item{p}{is a vector of weights}

\item{s}{is the scalar estimate in previous iteration}

\item{r}{is the rank of U0}
}
\description{
Perform an M-step update for estimating the scalar for prior covariance matrix U0
in the special case of V_j = I. U0 can be rank-deficient.

Perform an M-step update for estimating the scalar for prior covariance matrix U0
in the special case of V_j = I. U0 can be rank-deficient.
}
