% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{update_prior_covariance_unconstrained_fa}
\alias{update_prior_covariance_unconstrained_fa}
\title{Perform an M-step update for unconstrained prior covariance matrix U 
using factor analyzer in the special case of V_j = I. 
(See eq.(74) in main write-up)}
\usage{
update_prior_covariance_unconstrained_fa(X, U, p)
}
\arguments{
\item{X}{contains observed data of size n \times r.}

\item{U}{the estimate of U in previous iteration}

\item{p}{is a vector of weights}
}
\description{
Perform an M-step update for unconstrained prior covariance matrix U 
using factor analyzer in the special case of V_j = I. 
(See eq.(74) in main write-up)
}
