% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{simulate_ud_data}
\alias{simulate_ud_data}
\title{Simulate Data from Multivariate Normal Means Model}
\usage{
simulate_ud_data(n, w, U, V)
}
\arguments{
\item{n}{Number of data points to simulate.}

\item{w}{A numeric vector of length k specifying the prior mixture
weights. All entries must be non-negative, but need not sum to 1;
the mixture weights are automatically normalized to sum to 1.}

\item{U}{A list of length k specifying the covariance matrices in
the mixture-of-multivariate-normals prior; list element
\code{U[[i]]} is the m x m covariance matrix for the ith mixture
component.}

\item{V}{The m x m residual covariance matrix.}
}
\value{
An n x m matrix in which each row is a draw from the
  multivariate normal means model. For the univariate case (m = 1), a
  vector is returned.
}
\description{
Simulate data points from the multivariate normal
  means model. The univariate case (m = 1) is also handled. See
  \code{\link{ud_fit}} for the model definition.
}
\seealso{
\code{\link{ud_fit}}
}
