% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/posterior.R, R/updates.R
\name{ud_fit_em}
\alias{ud_fit_em}
\alias{compute_posterior_probs}
\alias{update_mixture_weights_em}
\alias{update_resid_covariance}
\alias{update_prior_covariances}
\title{Title Goes Here}
\usage{
ud_fit_em(
  fit,
  covupdates = rep("none", length(fit$U)),
  control = list(),
  verbose = TRUE
)

compute_posterior_probs(fit, version = c("Rcpp", "R"))

update_mixture_weights_em(fit, update = c("em", "none"))

update_resid_covariance(
  fit,
  update = c("em", "none"),
  version = c("Rcpp", "R")
)

update_prior_covariances(
  fit,
  covupdates = rep("none", length(fit$U)),
  minval = 1e-14
)
}
\arguments{
\item{fit}{Describe input argument "fit" here.}

\item{covupdates}{Describe input argument "covupdates" here.}

\item{control}{Describe input argumennt "control" here.}

\item{verbose}{Describe input argument "verbose" here.}

\item{version}{Describe input argument "version" here.}

\item{update}{Describe input argument "update" here.}

\item{minval}{Describe input argument "minval" here.}
}
\value{
Describe the outputs here.
}
\description{
Description of advanced fitting functions goes here.
}
\examples{
# Add examples here.

}
\keyword{internal}
