% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R, R/updates.R
\name{compute_posterior_probs}
\alias{compute_posterior_probs}
\alias{update_mixture_weights_em}
\alias{update_resid_covariance}
\alias{assign_prior_covariance_updates}
\alias{update_prior_covariances}
\title{Low-Level Interface for Fitting Ultimate Deconvolution Models}
\usage{
compute_posterior_probs(fit, version = c("Rcpp", "R"))

update_mixture_weights_em(fit, update = c("em", "none"))

update_resid_covariance(
  fit,
  update = c("em", "none"),
  version = c("Rcpp", "R")
)

assign_prior_covariance_updates(fit, control = list())

update_prior_covariances(
  fit,
  covupdates = assign_prior_covariance_updates(fit)$covupdates,
  minval = 1e-14
)
}
\arguments{
\item{fit}{An Ultimate Deconvolution model fit. Typically,
this will be an output of \code{\link{ud_init}} or \code{ud_fit}.}

\item{version}{Describe input argument "version" here.}

\item{update}{Describe input argument "update" here.}

\item{control}{A list of parameters controlling the behaviour of
the model fitting and initialization. See \code{\link{ud_fit}} for
details.}

\item{minval}{Describe input argument "minval" here.}

\item{covtypes}{Describe input argument "covtypes" here.}
}
\value{
An Ultimate Deconvolution model fit. See
  \code{\link{ud_fit}} for details.
}
\description{
Here we describe the low-level model fitting interface.
}
\details{
Add details here.
}
\examples{
# Add examples here.

}
\seealso{
\code{\link{ud_init}}, \code{\link{ud_fit}}
}
\keyword{internal}
