% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R, R/update_mixture_weights.R,
%   R/update_prior_covariances.R, R/update_resid_covariance.R
\name{compute_posterior_probs}
\alias{compute_posterior_probs}
\alias{update_mixture_weights}
\alias{assign_prior_covariance_updates}
\alias{update_prior_covariances}
\alias{update_resid_covariance}
\title{Low-Level Interface for Fitting Ultimate Deconvolution Models}
\usage{
compute_posterior_probs(fit, version = c("Rcpp", "R"))

update_mixture_weights(fit, update = c("em", "none"))

assign_prior_covariance_updates(fit, control = list())

update_prior_covariances(
  fit,
  covupdates = assign_prior_covariance_updates(fit)$covupdates,
  minval = 1e-14
)

update_resid_covariance(
  fit,
  update = c("em", "none"),
  version = c("Rcpp", "R")
)
}
\arguments{
\item{fit}{An Ultimate Deconvolution model fit. Typically,
this will be an output of \code{\link{ud_init}} or \code{ud_fit}.}

\item{version}{Describe input argument "version" here.}

\item{update}{Describe input argument "update" here.}

\item{control}{A list of parameters controlling the behaviour of
the model fitting and initialization. See \code{\link{ud_fit}} for
details.}

\item{minval}{Describe input argument "minval" here.}

\item{covtypes}{Describe input argument "covtypes" here.}
}
\value{
An Ultimate Deconvolution model fit. See
  \code{\link{ud_fit}} for details.
}
\description{
Here we describe the low-level model fitting interface.
}
\details{
Add details here.
}
\examples{
# Simulate data from a UD model.
set.seed(1)
n <- 4000
V <- rbind(c(0.8,0.2),
           c(0.2,1.5))
U <- list(none   = rbind(c(0,0),
                         c(0,0)),
          shared = rbind(c(1.0,0.9),
                         c(0.9,1.0)),
          only1  = rbind(c(1,0),
                         c(0,0)),
          only2  = rbind(c(0,0),
                         c(0,1)))
w <- c(0.8,0.1,0.075,0.025)
rownames(V) <- c("d1","d2")
colnames(V) <- c("d1","d2")
X <- simulate_ud_data(n,w,U,V)

# Perform 4 EM updates.
fit1 <- ud_init(X)
fit1 <- ud_fit(fit1,control = list(maxiter = 4))

# Update the responsibilities matrix.
fit2 <- compute_posterior_probs(fit1)

# Update the mixture weights.
fit2 <- update_mixture_weights(fit2)

# Update the residual covariance, V.
fit2 <- update_resid_covariance(fit2)

# Update the prior covariance matrices, U, using the defaults.
fit2 <- update_prior_covariances(fit2)

# Update only the unconstrained prior covariance matrices using the
# truncated eigenvalue decomposition ("ted") algorithm.
control <- list(scaled.update = "none",
                rank1.update = "none",
                unconstrained.update = "ted")
updates <- assign_prior_covariance_updates(fit,control)$covupdates
fit2 <- update_prior_covariances(fit2,updates)

# Compute the new log-likelihood and compare the old one.
print(logLik(fit1),digits = 8)
print(logLik(fit2),digits = 8)

}
\seealso{
\code{\link{ud_init}}, \code{\link{ud_fit}}
}
\keyword{internal}
