% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ud_init}
\alias{ud_init}
\title{Initialize Ultimate Deconvolution Model Fit}
\usage{
ud_init(
  X,
  V = diag(ncol(X)),
  n_rank1,
  n_unconstrained,
  U_scaled = list(indep = diag(ncol(X)), equal = matrix(1, ncol(X), ncol(X))),
  U_rank1,
  U_unconstrained,
  control = list()
)
}
\arguments{
\item{X}{An n x m data matrix, in which each row of the matrix is
an m-dimensional data point. \code{X} should have at least 2 rows
and 2 columns.}

\item{V}{Either an m x m matrix giving the initial estimate of the
residual covariance matrix, or a list of m x m covariance matrices
of length n.}

\item{n_rank1}{A non-negative integer specifying the number of
rank-1 covariance matrices included in the mixture prior. Initial
estimates of the m x m rank-1 covariance matrices are generated at
random. At most one of \code{n_rank1} and \code{U_rank1} should be
provided. If neither are specified, 4 random rank-1 matrices will be
included.}

\item{n_unconstrained}{A non-negative integer specifying the number
of unconstrained covariance matrices included in the mixture
prior. Initial estimates of the m x m covariance matrices are
generated at random. At most one of \code{n_unconstrained} and
\code{U_unconstrained} should be provided. If neither are
specified, 4 random unconstrained matrices will be included.}

\item{U_scaled}{A list specifying initial estimates of the scaled
covariance matrices in the mixture prior.}

\item{U_rank1}{A list specifying initial estimates of the rank-1
matrices in the mixture prior. At most one of \code{n_rank1} and
\code{U_rank1} should be provided. If \code{U_rank1} is not given,
the rank-1 covariates are initialized at random.}

\item{U_unconstrained}{A list specifying initial estimates of the
unconstrained matrices in the mixture prior. At most one of
\code{n_unconstrained} and \code{U_unconstrained} should be
provided. If \code{U_unconstrained} is not given, the matrices are
initialized at random.}

\item{control}{A list of parameters controlling the behaviour of
the model initialization. See \code{\link{ud_fit}} for details.}
}
\value{
An object of class "ud_fit". See \code{\link{ud_fit}} for
  details.
}
\description{
Initialize an Ultimate Deconvolution model fit. See
  \code{\link{ud_fit}} for background and model definition.
}
\seealso{
\code{\link{ud_fit}}
}
